the points mentioned in the comments have been fixed
