+++ src/main/java/ru/daniil/NauJava/entity/Report.java
@@ -31,25 +31,38 @@ public class Report {
     @Column(name = "execution_time_products")
     private Long executionTimeProducts;
 
+    /**
+     * Конструктор по умолчанию для формирования сущности отчёта.
+     * Указывает время создания сущности.
+     */
     public Report() {
         this.createdAt = LocalDateTime.now();
     }
 
+    /**
+     * Конструктор для формирования сущности отчёта с указанным статусом
+     * и временем создания.
+     */
     public Report(ReportStatus status) {
         this();
         this.status = status;
     }
 
     public Long getId() { return id; }
+
     public void setId(Long id) { this.id = id; }
 
     public ReportStatus getStatus() { return status; }
+
     public void setStatus(ReportStatus status) { this.status = status; }
 
     public String getContent() { return content; }
-    public void setContent(String content) { this.content = content; }
-    public void setCompletedAt(LocalDateTime completedAt) { this.completedAt = completedAt; }
-    public void setExecutionTimeUsers(Long executionTimeUsers) { this.executionTimeUsers = executionTimeUsers; }
-    public void setExecutionTimeProducts(Long executionTimeProducts) { this.executionTimeProducts = executionTimeProducts; }
 
+    public void setContent(String content) { this.content = content; }
+
+    public void setCompletedAt(LocalDateTime completedAt) { this.completedAt = completedAt; }
+
+    public void setExecutionTimeUsers(Long executionTimeUsers) { this.executionTimeUsers = executionTimeUsers; }
+
+    public void setExecutionTimeProducts(Long executionTimeProducts) { this.executionTimeProducts = executionTimeProducts; }
 }
